Refactor repeated components
