+++ ui/util/web.js
@@ -1,4 +1,4 @@
-const { URL, LBRY_WEB_STREAMING_API } = require('../../config');
+const { URL, LBRY_WEB_STREAMING_API, THUMBNAIL_CARDS_CDN_URL } = require('../../config');
 
 const CONTINENT_COOKIE = 'continent';
 
@@ -33,5 +33,27 @@ function generateDirectUrl(claimName, claimId) {
   return `${URL}/$/stream/${claimName}/${claimId}`;
 }
 
+function getThumbnailCdnUrl(url) {
+  if (
+    !THUMBNAIL_CARDS_CDN_URL ||
+    !url ||
+    (url && (url.includes('https://twitter-card') || url.includes('https://cards.odysee.com')))
+  ) {
+    return url;
+  }
+
+  if (url) {
+    const encodedURL = Buffer.from(url).toString('base64');
+    return `${THUMBNAIL_CARDS_CDN_URL}${encodedURL}.jpg`;
+  }
+}
+
 // module.exports needed since the web server imports this function
-module.exports = { generateStreamUrl, generateEmbedUrl, generateDownloadUrl, generateDirectUrl, CONTINENT_COOKIE };
+module.exports = {
+  generateStreamUrl,
+  generateEmbedUrl,
+  generateDownloadUrl,
+  generateDirectUrl,
+  getThumbnailCdnUrl,
+  CONTINENT_COOKIE,
+};
